5.6. Поправит библиотеку @project/blog-comment.
