+++ project/libs/blog/comments/src/module/comments.module.ts
@@ -4,10 +4,11 @@ import { CommentsController } from './comments.controller';
 import { CommentRepository } from './comment.repository';
 import { PrismaClientModule } from '@project/blog-models';
 import { CommentFactory } from './comment.factory';
+import { CommentsTechController } from './comments-tech.controller';
 
 @Module({
   imports: [PrismaClientModule],
-  controllers: [CommentsController],
+  controllers: [CommentsController, CommentsTechController],
   providers: [CommentsService, CommentRepository, CommentFactory],
   exports: [CommentsService],
 })
